added a merge_strategy parameter
added a merge_strategy parameter throughout the cascading code which allows 'obspy' style merging, or 'max', or 'both' which tries them in order. but not tested yet.
this will be important for when we do not want to blindly merge sds archives - for example, it is possible that some files have been instrument corrected in which case 'max' (or 'both') would be a disaster.